Move helper scripts location from helpers to bcnms folder
This way they will get cached.
