+++ scripts/zones/Apollyon/bcnms/ne_apollyon_helper.lua
