+++ scripts/zones/Apollyon/mobs/Ghost_Clot.lua
@@ -2,7 +2,7 @@
 -- Area: Apollyon SE, Floor 1
 --  Mob: Ghost Clot
 -----------------------------------
-require("scripts/zones/Apollyon/helpers/apollyon_se")
+require("scripts/zones/Apollyon/bcnms/se_apollyon_helper")
 -----------------------------------
 local entity = {}
