+++ scripts/zones/Apollyon/mobs/Kerkopes.lua
@@ -2,7 +2,7 @@
 -- Area: Apollyon NE, Floor 4
 --  Mob: Kerkopes
 -----------------------------------
-require("scripts/zones/Apollyon/helpers/apollyon_ne")
+require("scripts/zones/Apollyon/bcnms/ne_apollyon_helper")
 -----------------------------------
 local entity = {}
