+++ scripts/zones/Apollyon/mobs/Metalloid_Amoeba.lua
@@ -2,7 +2,7 @@
 -- Area: Apollyon SE, Floor 1
 --  Mob: Metalloid Amoeba
 -----------------------------------
-require("scripts/zones/Apollyon/helpers/apollyon_se")
+require("scripts/zones/Apollyon/bcnms/se_apollyon_helper")
 -----------------------------------
 local entity = {}
