+++ scripts/zones/Apollyon/mobs/Thiazi.lua
@@ -2,7 +2,7 @@
 -- Area: Apollyon NE, Floor 2
 --  Mob: Thiazi
 -----------------------------------
-require("scripts/zones/Apollyon/helpers/apollyon_ne")
+require("scripts/zones/Apollyon/bcnms/ne_apollyon_helper")
 -----------------------------------
 local entity = {}
