+++ scripts/zones/Apollyon/mobs/Troglodyte_Dhalmel.lua
@@ -3,7 +3,7 @@
 --  Mob: Troglodyte Dhalmel
 -----------------------------------
 local ID = require("scripts/zones/Apollyon/IDs")
-require("scripts/zones/Apollyon/helpers/apollyon_ne")
+require("scripts/zones/Apollyon/bcnms/ne_apollyon_helper")
 require("scripts/globals/pathfind")
 -----------------------------------
 local entity = {}
